Fix linting issues
This commit fixes various issues that occured due to
es-lint being added in a previous commit.

One such includes improperly calling this.setState,
and another was not having bindings functions to their
respective components - this was done the ES6 way, which
was to simply call the function in an anonymous function.

This also adds the prop-types package. prop-types will
be replaced by Flow types in a future PR.