Add Routes component with FragmentContainer
This commit puts the Routes div into its own component.

This was done as to wrap it in a Fragment container,
which ensures that its required data is loaded before
actually rendering it.

In addition, the query defined in the Routes_state
fragment container is now referred to in the root
query in App, which ensures that the Routes component
gets exactly what it wants as its query is defined in
one place.