12 - Restructure Relay Schema as to correspond to API Changes (#28)
Thanks to https://github.com/trynmaps/tryn-api/pull/14 the client
now has to change its Relay schema as to correspond to the new
format.

The reason such a change is needed is because it's necessary to
do the things we want, like only get data corresponding to open
routes or better handle the drawing of each route.

Now that the states are inside a route,
we can choose to request the vehicles ONLY for the routes that
are selected - which is a big performance gain. At the same
time, only a handful of routes will be open at a time, and there's
another PR that will let the user choose which ones
(https://github.com/trynmaps/tryn-react/pull/24).

However, while the client works - it only displays the state
for the first route. The next step is to have a layer for each
route, that contains all of its states - which is a TODO
described here: https://github.com/trynmaps/tryn-react/issues/27

Tests:
- manually opened web app and dragged around. Worked as expected.