34 - Update TrynState
This commit adds a DateTimePicker that allows us to choose a certain
time for the map state.

It also moves the trynState and input variables to the Map fragment
so that its RefetchContainer will actually update the props when
we perform a refetch, which is whenever the Date/Time gets changed.

https://github.com/facebook/relay/issues/2244