87 - add a new city + ui support to select it
Allows users to choose a different city, which changes the routes shown
in the control panel, changes the agency set in API requests, clears
routes/stops already drawn, moves the map, and uses a different geojson
file.

Cities are set in cities.json, and so new cities can now easily be
added. Note that the API currently only supports Muni and the
Toronto Transit Commission.